Created base fragments and placeholders
This includes a BrowseFragment that is ready to support images but
doesn't yet have an adapter. It also has a placeholder ViewModel that
can be expanded with more content. There are data models for Video and
VideoGroup, though we can consider better names.

The PlaybackFragment doesn't do anything at this point other than just
unparcel the passed Video. ExoPlayer is included via Gradle, so it
should be easy to start on that too.

This also includes some image helper classes that will blur and darken
images. They're intended for the backgrounds while browsing to ensure
images meet accessibility requirements. Right now they're using default
values, but we can tweak them once we have real content.

Change-Id: Id66efd7b4857983d6e8b34d6dec3b029605aa6ae