Fixed crash from issue #13 in Settings
All Fragments need to have a public default constructor, so that they
can be recreated by the system. In addition, they must be static if they
are inner classes because there will be no reference to the parent when
recreated.

I removed the Stack of Fragments that was being kept because this is
also a risky behavior as it's easy to keep references that should be
garbage collected. Fortunately, this sample doesn't need to retain these
Fragments to demo basic TV app settings.