Added placeholders for preferences
This uses the standard Leanback approach to preferences where there is a
LeanbackSettingsFragment(Compat) that serves as the root for all the
individual preference fragments. Right now, there's just one and it just
pops up a Toast when you interact with things, but that should make it
trivial to add others and handle interactions.

You can get to the preferences from the BrowseFragment using
BrowseFragmentDirections on the nav controller (similar to going to the
PlaybackFragment). Once the BrowseFragment has some content, we can put
a specific prefs item in it.

Change-Id: Ic2903d52b1538c9f226f451df16f5c9c45e096d1