Added support for playback watch progress
This uses Room to store the data in an extremely simple table. The table
just has a video ID (currently the video URL is used as the ID since it
is unique) and a start position. The start position comes directly from
ExoPlayer so that the video can be resumed from the same position.

The PlaybackFragment communicates with a PlaybackViewModel, which is
connected to a repository. The repository abstracts the data source, so
that the ViewModel doesn't know about future updates to store this data
in the cloud.

There is no handling for end-of-video at this point. If you finish a
video, it stops and that position will be saved. When you try to play
the video later, it will automatically seek to the end and not start
playback.

There is also a method for deleting all the watch progress, since that
is a coming feature, but it's not connected to anything at this point.

Change-Id: Ib3acbb0c09298986974bde8efb216275c2885f8c