Created a Firebase config option
This is a simple Gradle bool that is added to BuildConfig and checked at
runtime. If it's off, the user goes to a screen that explains the
limitations (placeholder text added for now) and has a button to
continue. If it's on, the user will be able to log in (functionality to
be added in a future commit).

Change-Id: I523f593d8035a98e8c21bf948272a8f2fb495dd4