Added video class properties and Moshi
Switched to parsing the JSON with Moshi to ensure that nonnull values
are respected in the Video class. This also adds an ID, URI,
Description, and VideoType to the Video class. All fields are non-null
and Moshi correctly enforces that.

There are also a few unit test updates related to this because the test
JSON needs to include all fields or it's invalid and Moshi will throw an
exception. Only the VideoParserUnitTest uses JSON now and the
BrowseViewModelUnitTest instead uses actual Video instances constructed
in the test. This should simplify additional changes to the Video data
model. The FileVideoRepositoryUnitTest was only really testing the
parsing, so it felt redundant and I removed it.

Change-Id: I229a5a715d40e85cbad91ed4be9d3ed42da89c05