Added initial support for MediaSession
This allows you to use the assistant to trigger the standard commands
like play, pause, seek, etc. Most of the work is done by the
MediaSessionConnector, which ties a MediaSession to the ExoPlayer. To be
more friendly to system resources, the app clears and recreates the
player and media session objects in onStop and onStart respectively.

Bug: 149105099
Change-Id: Ifdf212474f107dc46773293ad55dba45311289ec