Updated readme and added state machine readme
Added more project details to the main readme. This doesn't quite cover
everything, but we can add a little more once the last pieces of the
project fall into place. I also added a readme for the state machine.
It's focused on the high level concept and reasoning, relying on the
code itself to be the source of truth.

Change-Id: I2f7af933cf06c88c2a6e9027ae1a3827c9fd08ef