Removed jcenter and updated libraries
This removes the use of jcenter for libraries and swaps in maven
central. ExoPlayer 2.13.3 is the first release available outside of
jcenter, so updated to use that. This also updates leanback to the RC
(only change since our version is to SearchSupportFragment, which we're
not using) and navigation is now on 2.3.5, which appears to have only
had bug fixes that aren't relevant for this sample.

https://github.com/google/ExoPlayer/releases/tag/r2.13.3
https://developer.android.com/jetpack/androidx/releases/leanback#1.1.0-rc01
https://developer.android.com/jetpack/androidx/releases/navigation#2.3.5

Change-Id: I1d89c6a8aa100033fd4eabd8eb198ddcfe24cc76